Add option to slow qemu down
For slow targets and fast hosts, the emulation may be faster
than the actual hardware, which can be undesirable for various
reasons. Add a run-time option to slow down the emulation
by sleeping in the CPU emulation.

The option is not available when KVM or XEN acceleration is
enabled since it would defeat the purpose.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>